@@ -13,17 +13,16 @@
   "private": true,
   "dependencies": {
     "@angular-devkit/schematics": "^7.3.1",
-    "@angular/animations": "^7.2.4",
+    "@angular/animations": "^8.2.14",
     "@angular/cdk": "^7.3.3",
-    "@angular/common": "7.2.4",
-    "@angular/compiler": "7.2.4",
-    "@angular/core": "7.2.4",
-    "@angular/forms": "7.2.4",
-    "@angular/http": "7.2.4",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "8.2.14",
+    "@angular/forms": "8.2.14",
     "@angular/material": "^7.3.3",
-    "@angular/platform-browser": "7.2.4",
-    "@angular/platform-browser-dynamic": "7.2.4",
-    "@angular/router": "7.2.4",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/router": "8.2.14",
     "@ngrx/store": "^7.2.0",
     "chart.js": "^2.8.0",
     "core-js": "^2.4.1",
@@ -33,21 +32,21 @@
     "moment": "^2.24.0",
     "ng2-charts": "^2.2.3",
     "ngx-cookie-service": "^2.2.0",
-    "rxjs": "^6.4.0",
+    "rxjs": "^6.5.5",
     "ts-mockito": "^2.3.1",
     "tslib": "^1.11.1",
     "tslint-config-prettier": "^1.18.0",
-    "zone.js": "^0.8.29"
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "^7.3.1",
-    "@angular/compiler-cli": "7.2.4",
-    "@angular/language-service": "7.2.4",
+    "@angular-devkit/build-angular": "~0.803.25",
+    "@angular/cli": "^8.3.26",
+    "@angular/compiler-cli": "8.2.14",
+    "@angular/language-service": "8.2.14",
     "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
-    "codelyzer": "^4.0.1",
+    "codelyzer": "^5.0.1",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~2.0.0",
@@ -58,6 +57,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "3.2.4"
+    "typescript": "3.5.3"
   }
 }
