@@ -6,15 +6,15 @@ export const mainSiteRoutes: Routes = [
         path: '',
         component: MainSiteComponent,
         children: [
-            { path: '', loadChildren: './../../pages/main/main.page.module#MainPageModule' },
-            { path: 'news', loadChildren: './../../pages/main/main.page.module#MainPageModule' },
-            { path: 'rating', loadChildren: './../../pages/rating/rating.page.module#RatingPageModule' },
-            { path: 'archive', loadChildren: './../../pages/archive/archive.page.module#ArchivePageModule' },
-            { path: 'movies', loadChildren: './../../pages/movies/movies.page.module#MoviesPageModule' },
-            { path: 'rules', loadChildren: './../../pages/rules/rules.page.module#RulesPageModule' },
-            { path: 'teams', loadChildren: './../../pages/teams/teams.page.module#TeamsPageModule' },
-            { path: 'profile/:id', loadChildren: './../../pages/profile/profile.page.module#ProfilePageModule' },
-            { path: 'dfwc2019', loadChildren: './../../pages/dfwc2019/dfwc2019.page.module#Dfwc2019PageModule' },
+            { path: '', loadChildren: () => import('./../../pages/main/main.page.module').then(m => m.MainPageModule) },
+            { path: 'news', loadChildren: () => import('./../../pages/main/main.page.module').then(m => m.MainPageModule) },
+            { path: 'rating', loadChildren: () => import('./../../pages/rating/rating.page.module').then(m => m.RatingPageModule) },
+            { path: 'archive', loadChildren: () => import('./../../pages/archive/archive.page.module').then(m => m.ArchivePageModule) },
+            { path: 'movies', loadChildren: () => import('./../../pages/movies/movies.page.module').then(m => m.MoviesPageModule) },
+            { path: 'rules', loadChildren: () => import('./../../pages/rules/rules.page.module').then(m => m.RulesPageModule) },
+            { path: 'teams', loadChildren: () => import('./../../pages/teams/teams.page.module').then(m => m.TeamsPageModule) },
+            { path: 'profile/:id', loadChildren: () => import('./../../pages/profile/profile.page.module').then(m => m.ProfilePageModule) },
+            { path: 'dfwc2019', loadChildren: () => import('./../../pages/dfwc2019/dfwc2019.page.module').then(m => m.Dfwc2019PageModule) },
         ],
     },
 ];
